merkleTreeRoot-not working

+++ src/node.h
@@ -7,7 +7,7 @@
 
 //leaf node
 struct leaf_node {
-  uint64_t *hash;
+  size_t *hash;
   char* value;
 };
 typedef struct leaf_node LeafNode;
